Move analyzer_client.dart and dartdoc_client.dart
