+++ app/bin/tools/notify_service.dart
@@ -5,8 +5,8 @@
 import 'dart:async';
 
 import 'package:pub_dartlang_org/service/entrypoint/tools.dart';
-import 'package:pub_dartlang_org/shared/analyzer_client.dart';
-import 'package:pub_dartlang_org/shared/dartdoc_client.dart';
+import 'package:pub_dartlang_org/analyzer/analyzer_client.dart';
+import 'package:pub_dartlang_org/dartdoc/dartdoc_client.dart';
 import 'package:pub_dartlang_org/shared/search_client.dart';
 
 void _printHelp() {
