+++ app/lib/analyzer/pana_runner.dart
@@ -6,15 +6,16 @@ import 'dart:async';
 
 import 'package:logging/logging.dart';
 import 'package:pana/pana.dart';
+import 'package:pub_semver/pub_semver.dart';
 
 import '../job/job.dart';
 import '../scorecard/backend.dart';
 import '../scorecard/models.dart';
-import '../shared/analyzer_client.dart' show createPanaSummaryForLegacy;
 import '../shared/configuration.dart';
 import '../shared/packages_overrides.dart';
 import '../shared/platform.dart';
 import '../shared/tool_env.dart';
+import '../shared/urls.dart' as urls;
 
 final Logger _logger = Logger('pub.analyzer.pana');
 
@@ -204,4 +205,33 @@ PanaReport panaReportFromSummary(Summary summary, {List<String> flags}) {
     licenses: summary?.licenses,
     flags: flags,
   );
+}
+
+Summary createPanaSummaryForLegacy(String packageName, String packageVersion) {
+  return Summary(
+      runtimeInfo: PanaRuntimeInfo(),
+      packageName: packageName,
+      packageVersion: Version.parse(packageVersion),
+      pubspec: null,
+      pkgResolution: null,
+      dartFiles: null,
+      platform: null,
+      licenses: null,
+      health: null,
+      maintenance: null,
+      suggestions: <Suggestion>[
+        Suggestion.error(
+          'pubspec.sdk.legacy',
+          'Support Dart 2 in `pubspec.yaml`.',
+          'The SDK constraint in `pubspec.yaml` doesn\'t allow the Dart 2.0.0 release. '
+              'For information about upgrading it to be Dart 2 compatible, please see '
+              '[${urls.dartSiteRoot}/dart-2#migration](${urls.dartSiteRoot}/dart-2#migration).',
+        ),
+      ],
+      stats: Stats(
+        analyzeProcessElapsed: 0,
+        formatProcessElapsed: 0,
+        resolveProcessElapsed: 0,
+        totalElapsed: 0,
+      ));
 }
