+++ app/lib/frontend/templates/package_admin.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import '../../shared/analyzer_client.dart';
+import 'package:pub_dartlang_org/analyzer/analyzer_client.dart';
 import '../../shared/urls.dart' as urls;
 
 import '../models.dart';
