+++ app/lib/service/services.dart
@@ -6,7 +6,9 @@ import 'package:gcloud/storage.dart';
 
 import '../account/backend.dart';
 import '../account/consent_backend.dart';
+import '../analyzer/analyzer_client.dart';
 import '../dartdoc/backend.dart';
+import '../dartdoc/dartdoc_client.dart';
 import '../frontend/backend.dart';
 import '../frontend/email_sender.dart';
 import '../frontend/name_tracker.dart';
@@ -19,9 +21,7 @@ import '../scorecard/backend.dart';
 import '../search/backend.dart';
 import '../search/index_simple.dart';
 import '../search/updater.dart';
-import '../shared/analyzer_client.dart';
 import '../shared/configuration.dart';
-import '../shared/dartdoc_client.dart';
 import '../shared/popularity_storage.dart';
 import '../shared/redis_cache.dart' show withAppEngineAndCache;
 import '../shared/search_client.dart';
